dev: move analysis to testshared
