+++ test/testshared/analysis_test.go
@@ -1,4 +1,4 @@
-package test
+package testshared
 
 import (
 	"regexp"
